Reversion back to code before the UI board was immplemented.
The program was modified greatly to use the UI board. The source code
folder (along with the Git repository) was then copied to this folder
and the code reverted back to the old version so that copy would be
available for use with old machines.

A previous attempt was made to do this by branching back and then
merging the branch, but this FAILED (see the GIT history) because it
simply merged the old and the new and ended up at the same version.

Signed-off-by: Mike Schoonover <mike@skoonie.com>